+++ lib/ontologies_linked_data/models/concerns/submission_process.rb
@@ -2,8 +2,20 @@ module LinkedData
   module Concerns
     module SubmissionProcessable
 
-      def process_submission(logger, options={})
+      def process_submission(logger, options = {})
         LinkedData::Services::OntologyProcessor.new(self).process(logger, options)
+      end
+
+      def generate_missing_labels(logger)
+        LinkedData::Services::GenerateMissingLabels.new(self).process(logger, file_path: self.master_file_path)
+      end
+
+      def generate_obsolete_classes(logger)
+        LinkedData::Services::ObsoleteClassesGenerator.new(self).process(logger, file_path: self.master_file_path)
+      end
+
+      def extract_metadata(logger, options = {})
+        LinkedData::Services::SubmissionMetadataExtractor.new(self).process(logger, options)
       end
 
       def diff(logger, older)
@@ -14,7 +26,11 @@ module LinkedData
         LinkedData::Services::SubmissionDiffGenerator.new(self).process(logger)
       end
 
-      def index(logger, commit: true, optimize: true)
+      def index_all(logger, commit: true)
+        LinkedData::Services::OntologySubmissionAllDataIndexer.new(self).process(logger, commit: commit)
+      end
+
+      def index_terms(logger, commit: true, optimize: true)
         LinkedData::Services::OntologySubmissionIndexer.new(self).process(logger, commit: commit, optimize: optimize)
       end
 
@@ -22,8 +38,8 @@ module LinkedData
         LinkedData::Services::SubmissionPropertiesIndexer.new(self).process(logger, commit: commit, optimize: optimize)
       end
 
-      def archive
-        LinkedData::Services::OntologySubmissionArchiver.new(self ).process
+      def archive(force: false)
+        LinkedData::Services::OntologySubmissionArchiver.new(self).process(force: force)
       end
 
       def generate_rdf(logger, reasoning: true)
@@ -37,4 +53,3 @@ module LinkedData
     end
   end
 end
-
