+++ lib/ontologies_linked_data/models/properties/object_property.rb
@@ -34,6 +34,10 @@ module LinkedData
               LinkedData::Hypermedia::Link.new("ancestors", lambda {|m| "#{self.ontology_link(m)}/properties/#{CGI.escape(m.id.to_s)}/ancestors"}, self.uri_type),
               LinkedData::Hypermedia::Link.new("descendants", lambda {|m| "#{self.ontology_link(m)}/properties/#{CGI.escape(m.id.to_s)}/descendants"}, self.uri_type),
               LinkedData::Hypermedia::Link.new("tree", lambda {|m| "#{self.ontology_link(m)}/properties/#{CGI.escape(m.id.to_s)}/tree"}, self.uri_type)
+
+      enable_indexing(:prop_search_core1, :property) do |schema_generator|
+        index_schema(schema_generator)
+      end
     end
 
   end
