+++ lib/ontologies_linked_data/models/properties/ontology_property.rb
@@ -3,6 +3,36 @@ module LinkedData
   module Models
 
     class OntologyProperty < LinkedData::Models::Base
+      model :ontology_property, name_with: ->(p) { uuid_uri_generator(p) }
+
+      def self.index_schema(schema_generator)
+        schema_generator.add_field(:label, 'text_general', indexed: true, stored: true, multi_valued: true)
+        schema_generator.add_field(:labelGenerated, 'text_general', indexed: true, stored: true, multi_valued: true)
+
+        schema_generator.add_field(:definition, 'string', indexed: true, stored: true, multi_valued: true)
+        schema_generator.add_field(:submissionAcronym, 'string', indexed: true, stored: true, multi_valued: false)
+        schema_generator.add_field(:parents, 'string', indexed: true, stored: true, multi_valued: true)
+        schema_generator.add_field(:ontologyType, 'string', indexed: true, stored: true, multi_valued: false)
+        schema_generator.add_field(:propertyType, 'string', indexed: true, stored: true, multi_valued: false)
+        schema_generator.add_field(:ontologyId, 'string', indexed: true, stored: true, multi_valued: false)
+        schema_generator.add_field(:submissionId, 'pint', indexed: true, stored: true, multi_valued: false)
+
+        %i[label labelGenerated].each do |field|
+          schema_generator.add_copy_field(field, '_text_')
+          schema_generator.add_copy_field(field, "#{field}Exact")
+          schema_generator.add_copy_field(field, "#{field}Suggest")
+          schema_generator.add_copy_field(field, "#{field}SuggestEdge")
+          schema_generator.add_copy_field(field, "#{field}SuggestNgram")
+        end
+      end
+
+      enable_indexing(:prop_search_core1, :property) do |schema_generator|
+        index_schema(schema_generator)
+      end
+
+      def properties
+        self.unmapped
+      end
 
       def retrieve_ancestors
         retrieve_ancestors_descendants(:ancestors)
